refactored & cleaned the join building for compiled query with include
