+++ src/Marten/Marten.csproj
@@ -112,7 +112,10 @@
     <Compile Include="Linq\QueryHandlers\AggregateQueryHandler.cs" />
     <Compile Include="Linq\QueryHandlers\CachedQuery.cs" />
     <Compile Include="Linq\QueryHandlers\CachedQueryHandler.cs" />
-    <Compile Include="Linq\QueryHandlers\CompiledIncludeJoinBuilder.cs" />
+    <Compile Include="Linq\QueryHandlers\CompiledInclude\CompiledIncludeJoinBuilder.cs" />
+    <Compile Include="Linq\QueryHandlers\CompiledInclude\DefaultIncludeCallbackResolver.cs" />
+    <Compile Include="Linq\QueryHandlers\CompiledInclude\DictionaryIncludeCallbackResolver.cs" />
+    <Compile Include="Linq\QueryHandlers\CompiledInclude\IIncludeCallbackResolver.cs" />
     <Compile Include="Linq\QueryHandlers\EnumerableQueryHandler.cs" />
     <Compile Include="Linq\QueryHandlers\JsonQueryHandler.cs" />
     <Compile Include="Linq\QueryHandlers\ListQueryHandler.cs" />
